Added flag to return wiki to write mode
If backup execution fails in the middle, then the final toggle will not execute and the wiki will remain in read-only mode. Now, if you run the script with -t, it will not run the backup process, just put the wiki back into write mode. This can be set to run regularly or just a given period after the backup script should be completed in cron with backup.sh -t.